Converted BlockGenerator to ItemGenerator
Now we should have a good base for making it much easier to add
an item that the item generator knows how to handle.

At this point all instances of BlockGenerator have been replaced
with ItemGenerator. The only difference in the game being that
the blocks aren't currently colored. But we'll get there.