Fixed the row generation bug.
The bug was in the 'if/else if' comparisons for which pattern to
use. Also I wasn't actually using the percentage values correctly.
If the distribution was 10, 30, 60 (easy, medium, hard) it would
check:

if num < 10 { // easy }
else if num >= 10 && num < 30 { // medium }
else if num >= 60 { // hard }

That is super wrong. Need to start adding unit tests today for the
item generator and model to weed out any other silly bugs like
that.