Added views to store scene in Storyboard.
It has a button and handles the undo button click. Need to now add
the logic for undoing a turn and decreasing the user's currency.

Also need to add code to enable/disable it based on: 1) the user
having enough currency to purchase the item and 2) the user having
just used an "undo".