Fixed zposition bug with block counts.
I was adding the label to the block as a child node and they both
had zPositions. I think the zPosition of the child is added to the
zPosition of the parent, so the text on the blocks had a zPosition
of about 200 which is greater than the zPosition of any one item
in the game.