Different screens now have same number of blocks.
Next steps: 1) left alignment of blocks; they overlap with the
left wall; 2) game over/loss risk is kind of weird and seems
inconsistent; when blocks move down to ground they push the balls
through the ground then if you save yourself, the balls are below
the line of the ground and can't shoot; 3) to make it look better,
don't have ground and side walls filled in with gray. Instead,
have them display a line like the ceiling does.