Moved some ball functions into GameScene.
In the middle of transferring some common functionality into the
generic GameScene file to reduce complex code duplication. If
something works in one game scene class, it will work in the other.
If it's broken in one game scene, it'll be broken in the other.
Should make finding and fixing bugs easier.