Fixed up the code so it's back to normal.
Clear row reward now works as expected. I thought I was going to
need to change the way removing items was processed but we process
hits on all items first and then go through to remove items so we
shouldn't have any issues with is. All that's left to do is to blow
up items when a clear row reward is hit in the same row. ...That
actually might require a slight rewrite to processing hit items.