Fixed the bug when isPaused is auto-set to true.
Fixed the bug by adding a second variable named realPause.
Whenever isPaused is set, it checks to see whether or not
realPaused is the same value. If it is not, then the scene's
isPaused and the view's isPaused are both set to realPaused. The
only issue with this is that realPaused MUST be set prior to
isPaused.