Removed separate reward ad view controller.
That controller was needed months ago because when the reward ad
would dismiss, it would also dismiss the classic and levels view
controllers. I had no way of preventing that from happening, so I
needed to put the reward ad within its own view controller that it
could dismiss to fix it. But now it looks like Google fix that
issue.