Fixed some more bugs with this.
It mostly works now. The thing that still remains is when the game
lags (or some other random time), a ball will randomly fire in its
own weird direction. Now when this happens with the current method
we're using for generating ball nodes, the last ball will get
removed from the screen which leaves 0 balls on the screen. It
looks like this ball fires from the new origin point meaning one
of the balls that just landed is firing off randomly for some
reason.